requirements.txt
pyserial==3.4
crc16==0.1.1
